Move integration tests in `deserialize/mod.rs` to integration tests.
