+++ src/deserialize/mod.rs
@@ -409,282 +409,4 @@ pub fn from_edn<T: Deserialize>(edn: &Edn) -> Result<T, Error> {
 }
 
 #[cfg(test)]
-mod test {
-    use super::*;
-    #[cfg(feature = "sets")]
-    use crate::edn::Set;
-    use crate::edn::{List, Map, Vector};
-    use crate::{hmap, hset, map, set};
-
-    #[test]
-    fn unit() {
-        let nil = "nil";
-        let unit: () = from_str(nil).unwrap();
-
-        assert_eq!(unit, ());
-    }
-
-    #[test]
-    #[cfg(feature = "sets")]
-    fn deser_btreeset_with_error() {
-        let edn = "#{\"a\", 5, \"b\"}";
-        let err: Result<BTreeSet<u64>, Error> = from_str(edn);
-        assert_eq!(
-            err,
-            Err(Error::Deserialize(
-                "Cannot safely deserialize Set(Set({Str(\"a\"), Str(\"b\"), UInt(5)})) to BTreeSet"
-                    .to_string()
-            ))
-        );
-    }
-
-    #[test]
-    fn from_str_simple_vec() {
-        let edn = "[1 \"2\" 3.3 :b true \\c]";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Ok(Edn::Vector(Vector::new(vec![
-                Edn::UInt(1),
-                Edn::Str("2".to_string()),
-                Edn::Double(3.3.into()),
-                Edn::Key(":b".to_string()),
-                Edn::Bool(true),
-                Edn::Char('c')
-            ])))
-        );
-    }
-
-    #[test]
-    fn from_str_list_with_vec() {
-        let edn = "(1 \"2\" 3.3 :b [true \\c])";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Ok(Edn::List(List::new(vec![
-                Edn::UInt(1),
-                Edn::Str("2".to_string()),
-                Edn::Double(3.3.into()),
-                Edn::Key(":b".to_string()),
-                Edn::Vector(Vector::new(vec![Edn::Bool(true), Edn::Char('c')]))
-            ])))
-        );
-    }
-
-    #[test]
-    #[cfg(feature = "sets")]
-    fn from_str_list_with_set() {
-        let edn = "(1 -10 \"2\" 3.3 :b #{true \\c})";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Ok(Edn::List(List::new(vec![
-                Edn::UInt(1),
-                Edn::Int(-10),
-                Edn::Str("2".to_string()),
-                Edn::Double(3.3.into()),
-                Edn::Key(":b".to_string()),
-                Edn::Set(Set::new(set![Edn::Bool(true), Edn::Char('c')]))
-            ])))
-        );
-    }
-
-    #[test]
-    fn from_str_simple_map() {
-        let edn = "{:a \"2\" :b true :c nil}";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Ok(Edn::Map(Map::new(
-                map! {":a".to_string() => Edn::Str("2".to_string()),
-                ":b".to_string() => Edn::Bool(true), ":c".to_string() => Edn::Nil}
-            )))
-        );
-    }
-
-    #[test]
-    #[cfg(feature = "sets")]
-    fn from_str_complex_map() {
-        let edn = "{:a \"2\" :b [true false] :c #{:A {:a :b} nil}}";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Ok(Edn::Map(Map::new(map! {
-            ":a".to_string() =>Edn::Str("2".to_string()),
-            ":b".to_string() => Edn::Vector(Vector::new(vec![Edn::Bool(true), Edn::Bool(false)])),
-            ":c".to_string() => Edn::Set(Set::new(
-                set!{
-                    Edn::Map(Map::new(map!{":a".to_string() => Edn::Key(":b".to_string())})),
-                    Edn::Key(":A".to_string()),
-                    Edn::Nil}))})))
-        );
-    }
-
-    #[test]
-    fn from_str_wordy_str() {
-        let edn = "[\"hello brave new world\"]";
-
-        assert_eq!(
-            Edn::from_str(edn).unwrap(),
-            Edn::Vector(Vector::new(vec![Edn::Str(
-                "hello brave new world".to_string()
-            )]))
-        );
-    }
-    #[test]
-    fn uuid() {
-        let uuid = "#uuid \"af6d8699-f442-4dfd-8b26-37d80543186b\"";
-        let edn: Edn = Edn::from_str(uuid).unwrap();
-
-        assert_eq!(
-            edn,
-            Edn::Uuid("af6d8699-f442-4dfd-8b26-37d80543186b".to_string())
-        );
-    }
-
-    #[test]
-    fn deserialize_struct_with_vec() {
-        #[derive(PartialEq, Debug)]
-        struct Foo {
-            bar: Vec<Option<u64>>,
-        }
-        impl Deserialize for Foo {
-            fn deserialize(edn: &Edn) -> Result<Self, Error> {
-                Ok(Self {
-                    bar: from_edn(&edn[":bar"])?,
-                })
-            }
-        }
-        let edn_foo = "{:bar [1 nil 3]}";
-        let foo: Foo = from_str(edn_foo).unwrap();
-
-        assert_eq!(
-            foo,
-            Foo {
-                bar: vec![Some(1), None, Some(3)],
-            }
-        );
-    }
-
-    #[test]
-    fn test_sym() {
-        let edn: Edn = Edn::from_str("(a b c your-hair!-is+_parsed?)").unwrap();
-        let expected = Edn::List(List::new(vec![
-            Edn::Symbol("a".to_string()),
-            Edn::Symbol("b".to_string()),
-            Edn::Symbol("c".to_string()),
-            Edn::Symbol("your-hair!-is+_parsed?".to_string()),
-        ]));
-        assert_eq!(edn, expected);
-    }
-
-    #[test]
-    fn test_nft() {
-        let t: Edn = Edn::from_str("tTEST").unwrap();
-        let f: Edn = Edn::from_str("fTEST").unwrap();
-        let n: Edn = Edn::from_str("nTEST").unwrap();
-        let err: Edn = Edn::from_str("fTE").unwrap();
-
-        assert_eq!(n, Edn::Symbol("nTEST".to_string()));
-        assert_eq!(f, Edn::Symbol("fTEST".to_string()));
-        assert_eq!(t, Edn::Symbol("tTEST".to_string()));
-        assert_eq!(err, Edn::Symbol("fTE".to_string()));
-    }
-
-    #[test]
-    #[cfg(feature = "sets")]
-    fn test_more_sym() {
-        let edn: Edn = Edn::from_str("(a \\b \"c\" 5 #{hello world})").unwrap();
-        let expected = Edn::List(List::new(vec![
-            Edn::Symbol("a".to_string()),
-            Edn::Char('b'),
-            Edn::Str("c".to_string()),
-            Edn::UInt(5u64),
-            Edn::Set(Set::new(
-                set! { Edn::Symbol("hello".to_string()), Edn::Symbol("world".to_string()) },
-            )),
-        ]));
-        assert_eq!(edn, expected);
-    }
-
-    #[test]
-    fn deser_btreemap() {
-        let ns_map = Edn::Map(Map::new(map! {
-            ":a".to_string() => Edn::Vector(Vector::new(vec![Edn::Key(":val".to_string())])),
-            ":b".to_string() => Edn::Vector(Vector::new(vec![Edn::Key(":value".to_string())]))
-        }));
-        let expected = map! {
-            ":a".to_string() => vec![":val".to_string()],
-            ":b".to_string() => vec![":value".to_string()]
-        };
-        let map: std::collections::BTreeMap<String, Vec<String>> = from_edn(&ns_map).unwrap();
-        assert_eq!(map, expected);
-    }
-
-    #[test]
-    fn deser_hashmap() {
-        let ns_map = Edn::Map(Map::new(map! {
-            ":a".to_string() => Edn::Bool(true),
-            ":b".to_string() => Edn::Bool(false)
-        }));
-        let expected = hmap! {
-            ":a".to_string() => true,
-            ":b".to_string() => false
-        };
-        let map: std::collections::HashMap<String, bool> = from_edn(&ns_map).unwrap();
-        assert_eq!(map, expected);
-    }
-
-    #[test]
-    #[cfg(feature = "sets")]
-    fn deser_btreeset() {
-        let set = Edn::Set(Set::new(set! {
-            Edn::UInt(4),
-            Edn::UInt(5),
-            Edn::UInt(6)
-        }));
-        let expected = set! {
-            4,
-            5,
-            6,
-        };
-        let deser_set: std::collections::BTreeSet<u64> = from_edn(&set).unwrap();
-        assert_eq!(deser_set, expected);
-    }
-
-    #[cfg(feature = "sets")]
-    #[test]
-    fn deser_hashset() {
-        use ordered_float::OrderedFloat;
-
-        let set = Edn::Set(Set::new(set! {
-            Edn::Double(4.6.into()),
-            Edn::Double(5.6.into()),
-            Edn::Double(6.6.into())
-        }));
-        let expected = hset! {
-            OrderedFloat(4.6f64),
-            OrderedFloat(5.6f64),
-            OrderedFloat(6.6f64),
-        };
-        let deser_set: std::collections::HashSet<OrderedFloat<f64>> = from_edn(&set).unwrap();
-        assert_eq!(deser_set, expected);
-    }
-
-    #[test]
-    fn weird_input() {
-        let edn = "{:a]";
-
-        assert_eq!(
-            Edn::from_str(edn),
-            Err(Error::ParseEdn(
-                "Could not identify symbol index".to_string()
-            ))
-        );
-    }
-
-    #[test]
-    fn string_with_empty_set() {
-        assert_eq!("\"#{}\"", format!("{}", Edn::from_str("\"#{}\"").unwrap()));
-    }
-}
+mod test {}
